+++ spec/rubocop/cop/rspec/let_setup_spec.rb
@@ -116,6 +116,15 @@ RSpec.describe RuboCop::Cop::RSpec::LetSetup do
     RUBY
   end
 
+  it 'flags blockpass' do
+    expect_offense(<<-RUBY)
+      shared_context Foo do |&block|
+        let!(:bar, &block)
+        ^^^^^^^^^^^^^^^^^^ Do not use `let!` to setup objects not referenced in tests.
+      end
+    RUBY
+  end
+
   it 'complains when there is a custom nesting level' do
     expect_offense(<<-RUBY)
       describe Foo do