+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Fix `RSpec/EmptyExampleGroup` to flag example groups with examples in invalid scopes. ([@mlarraz][])
 * Fix `RSpec/EmptyExampleGroup` to ignore examples groups with examples defined inside iterators. ([@pirj][])
 * Improve `RSpec/NestedGroups`, `RSpec/FilePath`, `RSpec/DescribeMethod`, `RSpec/MultipleDescribes`, `RSpec/DescribeClass`'s top-level example group detection. ([@pirj][])
+* Add detection of `let!` with a block-pass or a string literal to `RSpec/LetSetup`. ([@pirj][])
 
 ## 1.42.0 (2020-07-09)